Initial

+++ src/cima/goes/img/_images.py
@@ -109,10 +109,10 @@ def get_clipped(image, image_region: LatLonRegion, clip: LatLonRegion):
     image_shape = image.shape
     pixels_per_lon = image_shape[0] / abs(image_region.lon_east-image_region.lon_west)
     pixels_per_lat = image_shape[1] / abs(image_region.lat_south-image_region.lat_north)
-    x = pixels_per_lon * abs(image_region.lon_east-clip.lon_west)
-    y = pixels_per_lat * abs(image_region.lat_north-clip.lat_north)
-    width = pixels_per_lon * abs(clip.lon_east-clip.lon_west)
-    height = pixels_per_lat * abs(clip.lat_south-clip.lat_north)
+    x = int(pixels_per_lon * abs(image_region.lon_east-clip.lon_west))
+    y = int(pixels_per_lat * abs(image_region.lat_north-clip.lat_north))
+    width = int(pixels_per_lon * abs(clip.lon_east-clip.lon_west))
+    height = int(pixels_per_lat * abs(clip.lat_south-clip.lat_north))
     print('CLIP:', x, y, width, height)
     return image[x:min(x+width, image_shape[0]), y:min(y+height, image_shape[1])]
